Refactored module /nistats/reporting.py in to /nistats/reporting/*.py
 - Preparation of incorporating Pypreprocess' GLM reporter.
